requirements.txt
Flask==0.9
Jinja2==2.6
WTForms==1.0.2
Werkzeug==0.8.3
argparse==1.2.1
beautifulsoup4==4.1.3
flask-peewee==0.6.1
peewee==2.0.3
python-dateutil==2.1
requests==0.14.2
six==1.2.0
wsgiref==0.1.2
wtf-peewee==0.2.0
